Move wasi-types-generated into wasi-types
