+++ Cargo.toml
@@ -44,8 +44,7 @@ members = [
     "lib/vm",
     "lib/wasi",
     "lib/wasi-types",
-    "lib/wasi-types-generated",
-    "lib/wasi-types-generated/wasi-types-generator-extra",
+    "lib/wasi-types/wasi-types-generator-extra",
     "lib/wasi-experimental-io-devices",
     "lib/wasi-local-networking",
     "lib/types",
