--- lib/wasi-types-generated/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "wasmer-wasi-types-generated"
-version = "3.0.0-beta"
-edition = "2018"
-license = "MIT"
-description = "Autogenerated wasi types, generated from .wit files"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-wasmer = { default-features = false, path = "../api", version = "3.0.0-beta" }
-wit-bindgen-rust = { package = "wasmer-wit-bindgen-rust", version = "0.1.1" }
-wit-bindgen-rust-wasm = { package = "wasmer-wit-bindgen-gen-rust-wasm", version = "0.1.1" }
-wit-bindgen-core = { package = "wasmer-wit-bindgen-gen-core", version = "0.1.1" }
-wit-parser = { package = "wasmer-wit-parser", version = "0.1.1" }
-
-[dev-dependencies.pretty_assertions]
-version = "1.3.0"
-
-[features]
-js = ["wasmer/js"]
-sys = ["wasmer/sys"]
