--- lib/wasi-types-generated/README.md
@@ -1,5 +1,0 @@
-Run `regenerate.sh` to regenerate the wasi-types from
-the `wasi-clean/typenames.wit` into the final Rust bindings.
-
-The `wasi-types-generator-extra` generates some extra code
-that wit-bindgen currently can't provide.
