+++ lib/wasi-types/Cargo.toml
@@ -10,15 +10,24 @@ license = "MIT"
 readme = "README.md"
 edition = "2018"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
+wasmer = { default-features = false, path = "../api", version = "3.0.0-beta" }
+wit-bindgen-rust = { package = "wasmer-wit-bindgen-rust", version = "0.1.1" }
+wit-bindgen-rust-wasm = { package = "wasmer-wit-bindgen-gen-rust-wasm", version = "0.1.1" }
+wit-bindgen-core = { package = "wasmer-wit-bindgen-gen-core", version = "0.1.1" }
+wit-parser = { package = "wasmer-wit-parser", version = "0.1.1" }
 wasmer-types = { path = "../types", version = "=3.0.0-beta" }
 wasmer-derive = { path = "../derive", version = "=3.0.0-beta" }
-wasmer-wasi-types-generated = { path = "../wasi-types-generated", version = "=3.0.0-beta" }
 serde = { version = "1.0", features = ["derive"], optional = true }
 byteorder = "1.3"
 time = "0.2"
 
+[dev-dependencies.pretty_assertions]
+version = "1.3.0"
+
 [features]
 enable-serde = ["serde", "wasmer-types/serde"]
-sys = ["wasmer-wasi-types-generated/sys"]
-js = ["wasmer-wasi-types-generated/js"]
+js = ["wasmer/js"]
+sys = ["wasmer/sys"]
