+++ lib/wasi-types/regenerate.sh
