+++ lib/wasi-types/src/wasi/bindings.rs
