+++ lib/wasi-types/src/wasi/extra_manual.rs
