+++ lib/wasi-types/wit-clean/output.wit
