+++ lib/wasi-types/wit-clean/typenames.wit
